Refine landing page
Landing page was acting as a list for all photos and photographers, but we only want a subset. Since we haven't implemented orders yet, we can't establish what is a 'popular' photo, so the listing of popular photos is really just a sampling. I also changed the All Photos view to use the existing photos partial since it was duplicating that code on its own page (with some slight differences).

Also changed the page title so it's not related to Dinner's Ready anymore, and I put in a link to the homepage in the navbar.