Add Download Center in user profile.
Early version of Download Center. Right now all photos for the user display as 100px squares and link to their dedicated photo pages where a user can see the large version (and might be able to download it if implemented in the future). I'd like to get checkboxes on each image as well as icons for quick download of individual images...