Fixed user_views_past_orders test
Renamed to user_visits_profile... currently testing for order history and failure to access if not logged in. NOT testing for Download History portion of profile page yet.

Altered users#show because tests like to run code that comes after a redirect even if it was redirected, and that blew stuff up. Now the show action skips over collection assignment if current_user is nil