Merge category and shop_category partials
Changed the calls to category and shop_category partials from sidebar and shop_sidebar partials to send in the needed path as a local variable. Using send method within category partial since the path is actually a helper method that takes a 'filter' parameter to filter images on the related pages.