Watermarks
watermark_id column added to stores table - the id is a photo_id, but I didn't want to make an official 'reference' to photo_id.
Changed the photo show page to use watermarks on the image preview on that page. It uses a method applied to the Store model, called watermark_accessor, that find the photo object using the watermark_id, then retrieves the appropriate 'public_id' that Cloudinary needs for pulling the actual watermark image. Retrieving that public_id is weird, though, because photo.file.public_id doesn't actually provide the public_id... you have to request photo.file.file.public_id to get the right one.
Next step is to build a watermark uploader on the admin page that uploads it like any other photo but then fills the store's watermark_id field with the created photo.id and sets that photo to be inactive so it doesn't appear on the site. I would also like to display the watermark on the admin page using a standard
placeholder image or simple black background.