Fix bug
Photo show calls watermark_accessor even when there is no watermark_id. This is normally OK since Cloudinary will just return the image without the watermark if it can't find the watermark (they call it an overlay) you provided. The problem with my code, however, was that the watermark_accessor starts by finding the correct photo. I was using find to get the photo, but find crashes if it doesn't find an object with the provided id or if it's given nil as the id. I changed the code to use find_by(:id) which returns nil instead of crashing.